requirements.txt
Flask
requests
Pillow
inference-sdk